fleet_management_service/requirements.txt
aio-pika==7.2.0
aiormq==6.2.3
anyio==3.5.0
asgiref==3.5.1
autopep8==1.6.0
click==8.1.3
fastapi==0.76.0
h11==0.13.0
httptools==0.4.0
idna==3.3
motor==3.0.0
multidict==6.0.2
pamqp==3.1.0
pika==1.2.1
pycodestyle==2.8.0
pydantic==1.9.0
pymongo==4.1.1
python-dotenv==0.20.0
PyYAML==6.0
sniffio==1.2.0
starlette==0.18.0
toml==0.10.2
typing_extensions==4.2.0
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
websockets==10.3
yarl==1.7.2
